Checkpoint moving over to new static/dynamic format
globals currently has the format

globals.arr.js -- runnable code
globals.arr.json -- static specification

These files are inspected and copied (currently, with no additional processing)
through the compiler to the compiled/builtin directory.

There are a number of TODOs, like:

- actually mirroring the directory structure with something like mkdir -p
- using a tool like Babel to do a JS-level module chase when saving the .arr.js
  files themselves

There's a key new concept, which is that when doing get-compiled(), a locator
can now respond with one of three options:

- This has already been compiled
- This needs to be compiled as a header + JS code
- This needs to be compiled as a Pyret module

Previously, the first two options were conflated, but we want to do some
processing of the JS modules in the future, so the second option needs to be
explicit.